Reorganize ArtScreen to improve readability
